Serve 410 page when png requested
